ETAColumn: -updated to show different icons for waiting time, reconnect waiting time ProgressColumn: fixed correct getValue DropboxCom: -fixed folder links -fixed regex IFolder: -updated folderregex -updated links regex -reenabled captcha method for ifolder AsFilecom: updated error handling Crocko: updated error handling, fixes #5496 NovaFile: fixed free download Rapidshare: fixed urls with %20 in it VeohCom: fixed for filesizes >2gb VipFileCom: -fixed premium download -added support for download mirrors -updated error handling WarServerCz: added support for md5sum
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@17810 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/plugins/hoster/VeohCom.java
@@ -178,7 +178,7 @@ public class VeohCom extends PluginForHost {
         if (fHash == null || fHash == null || sTime == null || videoId == null || fileSize == null || ext == null) { throw new PluginException(LinkStatus.ERROR_PLUGIN_DEFECT); }
 
         boolean oldDlHandling = fHash.length() < 32;
-        String hexTime = String.format("%08x", Integer.parseInt(sTime));
+        String hexTime = String.format("%08x", Long.parseLong(sTime));
 
         if (!oldDlHandling) {
             downloadLink.setName(downloadLink.getName() + ext);
@@ -203,10 +203,10 @@ public class VeohCom extends PluginForHost {
             sTime = br.getRegex("time=\'(\\d+)\'").getMatch(0);
             if (sTime == null) { throw new PluginException(LinkStatus.ERROR_PLUGIN_DEFECT); }
 
-            final String hexvidID = String.format("%08x", Integer.parseInt(videoId));
-            final String hexSize = String.format("%08x", Integer.parseInt(fileSize));
+            final String hexvidID = String.format("%08x", Long.parseLong(videoId));
+            final String hexSize = String.format("%08x", Long.parseLong(fileSize));
             final String templateUrl = br.getRegex("url base=\'(.*?)\'").getMatch(0);
-            hexTime = String.format("%08x", Integer.parseInt(sTime));
+            hexTime = String.format("%08x", Long.parseLong(sTime));
             String baseUrl = templateUrl;
 
             IV = fHash.substring(0, 16).getBytes();
